requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.13
botocore==1.37.13
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
dotenv==0.9.9
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
filelock==3.18.0
filetype==1.2.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.1
gotrue==2.11.4
greenlet==3.1.1
griffe==1.6.0
groq==0.19.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index-core==0.12.24.post1
llama-index-readers-file==0.4.7
llama-parse==0.6.4.post1
logfire==3.8.1
logfire-api==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistralai==1.5.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.66.3
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
packaging==24.2
pandas==2.2.3
pathtools==0.1.2
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.1
pydantic-ai==0.0.40
pydantic-ai-slim==0.0.40
pydantic-graph==0.0.40
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
pypdf==5.4.0
PyPDF2==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
sentry-sdk==2.24.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
storage3==0.11.3
streamlit==1.44.1
StrEnum==0.4.15
striprtf==0.0.26
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.8
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
